memf_symbols

Module btf

BTF (BPF Type Format) symbol resolver.

Parses the BTF section from a vmlinux binary or raw BTF data. BTF provides struct layouts and type information but NOT symbol addresses.

Reference: https://www.kernel.org/doc/html/latest/bpf/btf.html

Structs

BtfMember 🔒

BtfResolver

BTF symbol resolver.

BtfType 🔒

Enums

BtfKind 🔒

BTF type kinds.

Constants

BTF_HEADER_SIZE 🔒

BTF header size (version 1).

BTF_MAGIC 🔒

BTF header magic: 0xEB9F (little-endian).

Functions

extract_btf_from_elf 🔒

extract_btf_from_elf64 🔒

parse_type_section 🔒

read_btf_string 🔒

resolve_type_name 🔒